water-disease-predictor/backend/requirements.txt
flask==2.3.2
flask-CORS==4.0.0
pandas==2.0.3
numpy==1.26.4
matplotlib==3.8.0
scikit-learn==1.3.0
twilio==7.16.0
python-dotenv==1.0.0